+++ CHANGES.md
@@ -1,3 +1,5 @@
 # Changes
+## 0.2.0 (2019-12-11)
+- Redis: sentinels
 ## 0.1.0 (2019-11-20)
-- Initial release
+- Initial release: simple redis pool client